fix compiler warnings

+++ cpp/exe/icftool/ConsolePrinter.h
@@ -14,7 +14,7 @@ class ConsolePrinter : public ssp21::IMessagePrinter
 
     void write_line_prefix()
     {
-        for (int i = 0; i < indentation_level; ++i)
+        for (size_t i = 0; i < indentation_level; ++i)
         {
             std::cout << " ";
         }
